1-Checkmate/backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
tenacity
google-auth
